feat(#4) :: 디렉토리 변경
