--- src/main/java/baro/baro/domain/user/dto/req/req.java
@@ -1,4 +1,0 @@
-package baro.baro.domain.user.dto.req;
-
-public class req {
-}
