--- src/main/java/baro/baro/domain/user/dto/res/res.java
@@ -1,4 +1,0 @@
-package baro.baro.domain.user.dto.res;
-
-public class res {
-}
