+++ src/main/java/baro/baro/domain/user/service/AuthService.java
@@ -1,7 +1,7 @@
 package baro.baro.domain.user.service;
 
-import baro.baro.domain.user.entity.request.SignupRequest;
-import baro.baro.domain.user.entity.response.AuthTokensResponse;
+import baro.baro.domain.user.dto.req.SignupRequest;
+import baro.baro.domain.user.dto.res.AuthTokensResponse;
 
 public interface AuthService {
     AuthTokensResponse signup(SignupRequest request);
